pyproject.toml
[project]
name = "digital-media"
version = "0.1.0"
description = "Digital Media asset management service for Haderach"
requires-python = ">=3.12"
readme = "README.md"

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "firebase-admin>=6.4.0",
    "google-cloud-storage>=2.14.0",
    "google-cloud-aiplatform>=1.40.0",
    "google-cloud-vision>=3.5.0",
    "vertexai>=1.0.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "ruff>=0.2.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
